+++ NEWS.md
@@ -1,3 +1,4 @@
 # hellmer 0.1.0
 
 * Initial CRAN submission.
+* Removed `structured_data()` method as `texts()` now handles structured data responses.